Two essential fermion examples, and SS mkl support
Updated the two primary fermion example notebooks.

Also, added mkl support to all steadystate solvers.  when combined with
the c-code construction, this helps substantially in evaluating the
discrete boson + fermion leads example.

It can be a little unstable, so it maybe be better to have the option
off as default, even if someone has mkl installed. I am not sure.

I also fixed some missing transpose on returns in the ode solver in the
c-code, which had already been fixed in the python code.